requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.3
bleach==6.2.0
boto3==1.37.16
botocore==1.37.16
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci_sdr==0.0.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
ctc_segmentation==1.7.4
cycler==0.12.1
Cython==3.0.12
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
Distance==0.1.3
distro==1.9.0
docker==7.1.0
edge-tts==7.0.0
einops==0.8.1
elevenlabs==1.56.0
espnet==202205
espnet-tts-frontend==0.0.3
exceptiongroup==1.2.2
executing==2.2.0
fast_bss_eval==0.1.3
fastapi==0.115.11
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.0
g2p-en==2.1.0
groq==0.22.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h5py==3.13.0
halo==0.0.31
hiredis==3.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
indic-num2words==1.3.1
inflect==7.0.0
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kagglehub==0.3.11
kaldiio==2.18.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
litdata==0.2.32
litserve==0.2.6
llvmlite==0.44.0
log-symbols==0.0.14
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
opentelemetry-api==1.32.0
overrides==7.7.0
packaging==24.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pypinyin==0.44.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
pytorch-lightning==2.5.1
pytorch-wpe==0.0.1
pytz==2025.1
pyworld==0.3.5
PyYAML==6.0.2
pyzmq==26.3.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.11.4
Send2Trash==1.8.3
sentence-transformers==4.0.2
sentencepiece==0.2.0
simple-term-menu==1.6.6
six==1.17.0
smallestai==3.0.1
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spinners==0.0.24
srt==3.5.3
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
termcolor==3.0.1
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.5.1
torch-complex==0.4.4
torchaudio==2.5.1
torchmetrics==1.3.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.2
triton==3.1.0
typeguard==2.13.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
